# packages/polar/tests/fixtures/polar-2026-04-customers.openapi.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "Polar API",
    "version": "2026-04"
  },
  "paths": {
    "/v1/customers/": {
      "post": {
        "tags": [
          "customers",
          "public",
          "mcp",
          "cli"
        ],
        "summary": "Create Customer",
        "description": "Create a customer.\n\n**Scopes**: `customers:write`",
        "operationId": "customers:create",
        "security": [
          {
            "oidc": [
              "customers:write"
            ]
          },
          {
            "pat": [
              "customers:write"
            ]
          },
          {
            "oat": [
              "customers:write"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-polar-allowed-subjects": [
          "Organization",
          "User"
        ],
        "x-speakeasy-group": "customers",
        "x-speakeasy-name-override": "create",
        "x-codeSamples": [
          {
            "lang": "python",
            "source": "from polar.v2026_04 import Polar\n\npolar = Polar(\"polar_oat_xxx\")\n\nresponse = polar.customers.create(\n    external_id='usr_1337',\n    type='individual',\n    email='customer@example.com',\n)\nprint(response)\n"
          },
          {
            "lang": "typescript",
            "source": "import { createPolar } from \"@polar-sh/sdk/2026-04\";\n\nconst polar = createPolar({\n  accessToken: \"polar_oat_xxx\",\n});\n\nconst response = await polar.customers.create(\n  {\n    \"external_id\": \"usr_1337\",\n    \"type\": \"individual\",\n    \"email\": \"customer@example.com\"\n  },\n);\nconsole.log(response);\n"
          }
        ]
      }
    },
    "/v1/customers/{id}": {
      "get": {
        "tags": [
          "customers",
          "public",
          "mcp",
          "cli"
        ],
        "summary": "Get Customer",
        "description": "Get a customer by ID.\n\n**Scopes**: `customers:read` `customers:write`",
        "operationId": "customers:get",
        "security": [
          {
            "oidc": [
              "customers:read",
              "customers:write"
            ]
          },
          {
            "pat": [
              "customers:read",
              "customers:write"
            ]
          },
          {
            "oat": [
              "customers:read",
              "customers:write"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid4",
              "description": "The customer ID.",
              "title": "Id"
            },
            "description": "The customer ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFound"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-polar-allowed-subjects": [
          "Organization",
          "User"
        ],
        "x-speakeasy-group": "customers",
        "x-speakeasy-name-override": "get",
        "x-codeSamples": [
          {
            "lang": "python",
            "source": "from polar.v2026_04 import Polar\n\npolar = Polar(\"polar_oat_xxx\")\n\nresponse = polar.customers.get(\n    '00000000-0000-4000-8000-000000000000',\n)\nprint(response)\n"
          },
          {
            "lang": "typescript",
            "source": "import { createPolar } from \"@polar-sh/sdk/2026-04\";\n\nconst polar = createPolar({\n  accessToken: \"polar_oat_xxx\",\n});\n\nconst response = await polar.customers.get(\n  \"00000000-0000-4000-8000-000000000000\",\n);\nconsole.log(response);\n"
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "oat": {
        "type": "http",
        "description": "You can generate an **Organization Access Token** from your organization's settings.",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Customer": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/CustomerIndividual"
          },
          {
            "$ref": "#/components/schemas/CustomerTeam"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "individual": "#/components/schemas/CustomerIndividual",
            "team": "#/components/schemas/CustomerTeam"
          }
        }
      },
      "CustomerCreate": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/CustomerIndividualCreate"
          },
          {
            "$ref": "#/components/schemas/CustomerTeamCreate"
          }
        ]
      },
      "CustomerIndividual": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id",
            "description": "The ID of the customer.",
            "examples": [
              "992fae2a-2a17-4b7a-8d9e-e287cf90131b"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp of the object.",
            "examples": [
              "2026-01-01T00:00:00.000000Z"
            ]
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "examples": [
                  "2026-01-01T00:00:00.000000Z"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At",
            "description": "Last modification timestamp of the object."
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataOutputType"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id",
            "description": "The ID of the customer in your system. This must be unique within the organization. Once set, it can't be updated.",
            "examples": [
              "usr_1337"
            ]
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "The email address of the customer. This must be unique within the organization.",
            "examples": [
              "customer@example.com"
            ]
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified",
            "description": "Whether the customer email address is verified. The address is automatically verified when the customer accesses the customer portal using their email address.",
            "examples": [
              true
            ]
          },
          "type": {
            "type": "string",
            "const": "individual",
            "title": "Type",
            "description": "The type of customer.",
            "examples": [
              "individual"
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The name of the customer.",
            "examples": [
              "John Doe"
            ]
          },
          "billing_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Name",
            "description": "The name that should appear on the customer's invoices. Falls back to the customer name when not explicitly set.",
            "examples": [
              "John Doe"
            ]
          },
          "billing_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "tax_id": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "string"
                  },
                  {
                    "$ref": "#/components/schemas/TaxIDFormat"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2,
                "examples": [
                  [
                    "911144442",
                    "us_ein"
                  ],
                  [
                    "FR61954506077",
                    "eu_vat"
                  ]
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Organization Id",
            "description": "The ID of the organization owning the customer.",
            "examples": [
              "1dbfc517-0bbf-4301-9ba8-555ca42b9737"
            ]
          },
          "default_payment_method_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Payment Method Id",
            "description": "The ID of the customer's default payment method, if any. Use the payment methods endpoint to retrieve its details."
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "examples": [
                  "2026-01-01T00:00:00.000000Z"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "Timestamp for when the customer was soft deleted."
          },
          "first_user_event_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "examples": [
                  "2026-01-01T00:00:00.000000Z"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "First User Event At",
            "description": "Timestamp of the first event ingested for this customer. Can predate `created_at`, and is null if no event was ever ingested."
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "examples": [
              "https://www.gravatar.com/avatar/xxx?d=404"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "modified_at",
          "metadata",
          "email",
          "email_verified",
          "type",
          "name",
          "billing_name",
          "billing_address",
          "tax_id",
          "organization_id",
          "deleted_at",
          "first_user_event_at",
          "avatar_url"
        ],
        "title": "CustomerIndividual",
        "description": "A customer in an organization."
      },
      "CustomerIndividualCreate": {
        "properties": {
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 500,
                  "minLength": 1
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "propertyNames": {
              "maxLength": 40,
              "minLength": 1
            },
            "type": "object",
            "maxProperties": 50,
            "title": "Metadata",
            "description": "Key-value object allowing you to store additional information.\n\nThe key must be a string with a maximum length of **40 characters**.\nThe value must be either:\n\n* A string with a maximum length of **500 characters**\n* An integer\n* A floating-point number\n* A boolean\n\nYou can store up to **50 key-value pairs**."
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id",
            "description": "The ID of the customer in your system. This must be unique within the organization. Once set, it can't be updated.",
            "examples": [
              "usr_1337"
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "description": "The name of the customer.",
                "examples": [
                  "John Doe"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "billing_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AddressInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "tax_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "description": "Locale of the customer, given as an IETF BCP 47 language tag, e.g. `en`, `en-US` or `en-GB-oxendict`. If `null` or unsupported, the locale will default to `en`.",
                "examples": [
                  "en",
                  "en-US",
                  "fr",
                  "fr-CA"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4",
                "description": "The organization ID.",
                "examples": [
                  "1dbfc517-0bbf-4301-9ba8-555ca42b9737"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id",
            "description": "The ID of the organization owning the customer. **Required unless you use an organization token.**"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemberOwnerCreate"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional owner member to create with the customer. If not provided, an owner member will be automatically created using the customer's email and name."
          },
          "type": {
            "type": "string",
            "const": "individual",
            "title": "Type",
            "default": "individual"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "The email address of the customer. This must be unique within the organization.",
            "examples": [
              "customer@example.com"
            ]
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "CustomerIndividualCreate"
      },
      "CustomerTeam": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id",
            "description": "The ID of the customer.",
            "examples": [
              "992fae2a-2a17-4b7a-8d9e-e287cf90131b"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp of the object.",
            "examples": [
              "2026-01-01T00:00:00.000000Z"
            ]
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "examples": [
                  "2026-01-01T00:00:00.000000Z"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At",
            "description": "Last modification timestamp of the object."
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataOutputType"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id",
            "description": "The ID of the customer in your system. This must be unique within the organization. Once set, it can't be updated.",
            "examples": [
              "usr_1337"
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "The email address of the customer. This must be unique within the organization.",
            "examples": [
              "customer@example.com"
            ]
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified",
            "description": "Whether the customer email address is verified. The address is automatically verified when the customer accesses the customer portal using their email address.",
            "examples": [
              true
            ]
          },
          "type": {
            "type": "string",
            "const": "team",
            "title": "Type",
            "description": "The type of customer. Team customers can have multiple members.",
            "examples": [
              "team"
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "The name of the customer.",
            "examples": [
              "John Doe"
            ]
          },
          "billing_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Name",
            "description": "The name that should appear on the customer's invoices. Falls back to the customer name when not explicitly set.",
            "examples": [
              "John Doe"
            ]
          },
          "billing_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "tax_id": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "string"
                  },
                  {
                    "$ref": "#/components/schemas/TaxIDFormat"
                  }
                ],
                "type": "array",
                "maxItems": 2,
                "minItems": 2,
                "examples": [
                  [
                    "911144442",
                    "us_ein"
                  ],
                  [
                    "FR61954506077",
                    "eu_vat"
                  ]
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Organization Id",
            "description": "The ID of the organization owning the customer.",
            "examples": [
              "1dbfc517-0bbf-4301-9ba8-555ca42b9737"
            ]
          },
          "default_payment_method_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Payment Method Id",
            "description": "The ID of the customer's default payment method, if any. Use the payment methods endpoint to retrieve its details."
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "examples": [
                  "2026-01-01T00:00:00.000000Z"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "Timestamp for when the customer was soft deleted."
          },
          "first_user_event_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time",
                "examples": [
                  "2026-01-01T00:00:00.000000Z"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "First User Event At",
            "description": "Timestamp of the first event ingested for this customer. Can predate `created_at`, and is null if no event was ever ingested."
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url",
            "examples": [
              "https://www.gravatar.com/avatar/xxx?d=404"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "modified_at",
          "metadata",
          "email_verified",
          "type",
          "name",
          "billing_name",
          "billing_address",
          "tax_id",
          "organization_id",
          "deleted_at",
          "first_user_event_at",
          "avatar_url"
        ],
        "title": "CustomerTeam",
        "description": "A team customer in an organization."
      },
      "CustomerTeamCreate": {
        "properties": {
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 500,
                  "minLength": 1
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "propertyNames": {
              "maxLength": 40,
              "minLength": 1
            },
            "type": "object",
            "maxProperties": 50,
            "title": "Metadata",
            "description": "Key-value object allowing you to store additional information.\n\nThe key must be a string with a maximum length of **40 characters**.\nThe value must be either:\n\n* A string with a maximum length of **500 characters**\n* An integer\n* A floating-point number\n* A boolean\n\nYou can store up to **50 key-value pairs**."
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id",
            "description": "The ID of the customer in your system. This must be unique within the organization. Once set, it can't be updated.",
            "examples": [
              "usr_1337"
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "description": "The name of the customer.",
                "examples": [
                  "John Doe"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "billing_address": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AddressInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "tax_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "description": "Locale of the customer, given as an IETF BCP 47 language tag, e.g. `en`, `en-US` or `en-GB-oxendict`. If `null` or unsupported, the locale will default to `en`.",
                "examples": [
                  "en",
                  "en-US",
                  "fr",
                  "fr-CA"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid4",
                "description": "The organization ID.",
                "examples": [
                  "1dbfc517-0bbf-4301-9ba8-555ca42b9737"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id",
            "description": "The ID of the organization owning the customer. **Required unless you use an organization token.**"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemberOwnerCreate"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional owner member to create with the customer. If not provided, an owner member will be automatically created using the customer's email and name."
          },
          "type": {
            "type": "string",
            "const": "team",
            "title": "Type"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "The email address of the team customer. Optional for team customers — if omitted, an owner with an email must be provided.",
            "examples": [
              "customer@example.com"
            ]
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "CustomerTeamCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResourceNotFound": {
        "properties": {
          "error": {
            "type": "string",
            "const": "ResourceNotFound",
            "title": "Error",
            "examples": [
              "ResourceNotFound"
            ]
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "error",
          "detail"
        ],
        "title": "ResourceNotFound"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}